% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_auditc_score}
\alias{compute_auditc_score}
\title{Compute the AUDIT-C score}
\usage{
compute_auditc_score(
  x,
  audit_coefficients = default_coefficients,
  zero_if_never_drinks = FALSE
)
}
\arguments{
\item{x}{a data frame or matrix with three character columns containing, in order: \describe{
   \item{audit1_label}{a value in \code{'Never'}, \code{'Monthly or less'},
                       \code{'2 to 4 times a month'}, \code{'2 to 3 times a week'},
                       \code{'4 to 5 times a week'}, \code{'6 or more times a week'}}
   \item{audit2_label}{a value in \code{"1 to 2"}, \code{"3 to 4"}, \code{"5 to 6"},
                       \code{"7 to 9"}, \code{"10 to 12"}, \code{"13 to 15"}, \code{"16 or more"}}
   \item{audit3_label}{a value in \code{"Never"}, \code{"Less than monthly"}, \code{"Monthly"},
                       \code{"Weekly"}, \code{"Daily or almost daily"}}
}}

\item{audit_coefficients}{a list of coefficients (default: \code{\link[EWAC]{default_coefficients}})}

\item{zero_if_never_drinks}{a boolean indicating whether an AUDIT-C score of zero should be
given by default if AUDIT-1 = 'Never'. Default is FALSE.}
}
\value{
a vector of AUDIT-C score between 0 and 12.
}
\description{
Compute the traditional AUDIT-C score from Extended AUDIT-C data
}
\examples{
compute_auditc_score(data.frame(list(
   audit1_label = "2 to 3 times a week",
   audit2_label = "1 to 2",
   audit3_label = "Monthly"
)))
}
