% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_coefficients.R
\docType{data}
\name{default_coefficients}
\alias{default_coefficients}
\title{Default Extended AUDIT-C coefficients}
\format{
An object of class \code{list} of length 4.
}
\usage{
default_coefficients
}
\description{
A list of coefficients to
}
\details{
The list contains four elements:
\describe{
  \item{audit1}{a data frame for the AUDIT-1 question, containing the response item labels,
  EWAC weekly usual drinking frequency value, and AUDIT-C score.}
  \item{audit2}{a data frame for the AUDIT-2 question, containing the response item labels,
  EWAC quantity of alcohol value (in UK units per week), and AUDIT-C score.}
  \item{audit3}{a data frame for the AUDIT-3 question, containing the response item labels,
  EWAC variability of intense drinking value (weekly frequency), and AUDIT-C score.}
  \item{binge_value}{a numeric value corresponding to the average alcohol UK units consumed
  on an intense drinking occasion}
}
}
\keyword{data}
