% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWAC-package.R
\docType{package}
\name{EWAC-package}
\alias{EWAC}
\alias{EWAC-package}
\title{EWAC: R Package to Compute an Estimator of Weekly Alcohol Consumption (EWAC) from the Extended AUDIT-C}
\description{
>
  Compute an Estimator of Weekly Alcohol Consumption (EWAC) 
  from the Extended Alcohol Use Disorders Identification Tool.
}
\details{
R Package to Compute an Estimator of Weekly Alcohol Consumption (EWAC) from the Extended AUDIT-C

   Developed by \insertCite{DuteyMagni2020;textual}{EWAC}.

   Copyright (C) 2020 University of Southampton

   This program is free software; you can redistribute it and/or modify
   it under the terms of the GNU General Public License as published by
   the Free Software Foundation; either version 3 of the License, or
   (at your option) any later version.

   This program is distributed in the hope that it will be useful,
   but WITHOUT ANY WARRANTY; without even the implied warranty of
   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
   GNU General Public License for more details.
}
\references{
\insertAllCited{}
}
\seealso{
EWAC web app: \url{https://ewac.netlify.app/}
}
\author{
\strong{Maintainer}: Peter Dutey-Magni \email{p.dutey-magni@ucl.ac.uk} (\href{https://orcid.org/0000-0002-8942-9836}{ORCID})

Other contributors:
\itemize{
  \item University of Southampton [copyright holder]
}

}
\keyword{internal}
