% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_ewac.R
\docType{data}
\name{example_ewac}
\alias{example_ewac}
\title{Example dataset}
\format{
An object of class \code{data.frame} with 10 rows and 4 columns.
}
\usage{
example_ewac
}
\description{
A dataset consisting of 10 example Extended AUDIT-C values.
}
\details{
A data frame of 10 rows with teh following variables:
\describe{
  \item{person_id}{a unique identifier for the observation}
  \item{audit1_label}{character vector of AUDIT-1 answers (frequency)}
  \item{audit2_label}{character vector of AUDIT-2 answers (quantity)}
  \item{audit3_label}{character vector of AUDIT-3 answers (intense drinking)}
}
}
\keyword{data}
